Fixed models when there are duplicate values within a foregin model
Fixed models when there are duplicate values within a foregin model,
although the way I unique-ified the not-foreign model seemed very
hacky.  It seems to resemble what relational databases do when you
group by and ask for a non aggregated column… but you're not supposed
to do that for a reason.